move IProjectOptions interface
